refactor graph folder structure
